requirements.txt
aiogram==2.14.3
airtable-python-wrapper==0.15.3
emoji==1.4.2
numpy==1.21.1
pandas
schedule
